+++ Unity/Assets/Scripts/Model/Share/LockStep/IRoomMessage.cs
@@ -2,6 +2,6 @@ namespace ET
 {
     public interface IRoomMessage: IMessage
     {
-        int SeatIndex { get; set; }
+        byte SeatIndex { get; set; }
     }
 }
